Fix #1 Level 1 Heading is Title only on first line
The script was confusing comments inside source code inside markdown beginning
with '#' for Level 1 Markdown headings (that also begin with a '#', and hence
using the comment for the Title of the document.

Now we treat a Level 1 heading (beginning with a #) as the Title of the
document only if it is on the first line.